Refactor: update menu + pending donation page
Still missing history, popup
