+++ app/components/DonationPage/AddNewItem.jsx
@@ -1,20 +1,33 @@
 import React from 'react';
 import TextField from 'material-ui/TextField';
 
+const styles = {
+  textField: {
+    width: '100%',
+    maxWidth: 500,
+    height: 55,
+    marginRight: 15,
+    lineHeight: 2.7,
+    paddingBottom: 30,
+    letterSpacing: 1.9,
+    fontFamily: "'Open Sans', sans-serif"
+  }
+};
+
 const AddNewItem = ({ onUpdateItem, enableAddItem, onAddItem }) => (
-  <div className="text-flex">
+  <form>
     <TextField
       type="text"
       onKeyUp={onUpdateItem}
       id="donationTitle"
       hintText="What would you like to donate ?"
-      style={{ width: '100%', height: 55, marginRight: 45, lineHeight: 1.7, letterSpacing: 1.9, fontFamily: "'Open Sans', sans-serif", fontSize: 20 }}
+      style={styles.textField}
     />
     <button
-      className={enableAddItem ? 'btn-donation' : 'btn-donation btn-disabled'}
+      className={enableAddItem ? 'pointer-cursor btn-addDonation' : 'pointer-cursor btn-addDonation btn-disabled'}
       onClick={onAddItem}
     />
-  </div>
+  </form>
 );
 
 AddNewItem.propTypes = {
