+++ app/components/Reusable/TruckButton.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-const TruckButton = () => (<div className="truck-button"></div>);
+const TruckButton = () => (<div className="truck"></div>);
 
 module.exports = TruckButton;
