+++ app/stylesheets/components/footer.css
@@ -6,4 +6,7 @@
 	a:hover {
 		text-decoration: underline;
 	}
+  a:visited {
+    color: inherit;
+  }
 }
