+++ app/stylesheets/utils/text.css
@@ -25,12 +25,3 @@
 .text-right {
 	text-align: right;
 }
-.text-flex {
-	display: flex;
-}
-.fontSize-small {
-	font-size: 13px;
-}
-.fontSize-medium {
-	font-size: 16px;
-}
